% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entangler.R
\name{entangler}
\alias{entangler}
\title{EntangledClone game for R}
\usage{
entangler(width = NULL, height = NULL)
}
\arguments{
\item{width}{\code{integer} width in pixels}

\item{height}{\code{integer} height in pixels}
}
\value{
\code{htmlwidget} to be used in the console, RStudio viewer, browser, \code{rmarkdown},
   or Rstudio Shiny.
}
\description{
\href{https://github.com/Cacodaimon/EntangledClone}{EntangledClone} is a
   \code{d3.js} based Entanglement clone written in TypeScript.  Since it is
   JS/CSS, we can play it in R with this function.
}

