% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\docType{data}
\name{gameR_colors}
\alias{gameR_colors}
\title{List of discrete colours}
\format{
An object of class \code{list} of length 9.
}
\usage{
gameR_colors
}
\description{
List of discrete colours
}
\keyword{datasets}
