% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_tetris.R
\name{play_tetris}
\alias{play_tetris}
\title{Tetris Game}
\usage{
play_tetris(nhint = 5, speed = 1, color, bg = "white", ...)
}
\arguments{
\item{nhint}{the number of hint squares on the right.}

\item{speed}{the level of speed at which block is falling.}

\item{color}{colors of seven tetrominoes. Default color set is generated by \code{\link[grDevices]{hcl.colors}} with the palette "Dark 3".}

\item{bg}{the color to be used for the background of the device region.}

\item{...}{optional arguments to \code{\link[grDevices]{x11}}, except for width and height.}
}
\description{
This function is an implementation of Tetris based on X window system.
}
\examples{
\dontrun{
play_tetris()
}

}
