% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entangler.R
\name{renderEntangler}
\alias{renderEntangler}
\title{Widget render function for use in Shiny}
\usage{
renderEntangler(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

