% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner.R
\name{spinner}
\alias{spinner}
\title{Simple spinner animation}
\usage{
spinner(groups = 3, nframes = 100, fps = 10)
}
\arguments{
\item{groups}{number of groups}

\item{nframes}{number of frames}

\item{fps}{frames per second}
}
\description{
Simple spinner animation
}
