% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{gameR_colors}
\alias{gameR_colors}
\title{List of palettes}
\format{
An object of class \code{list} of length 13.
}
\usage{
gameR_colors
}
\arguments{
\item{palette}{Character name of palette. Either
blocks, cowboy, cups, cyberpunk, fallout, gris, ocarina, p4g, rayman, spirit, splat, superbros, wow)`}
}
\description{
List of palettes
}
\keyword{datasets}
