% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{gameR_cols}
\alias{gameR_cols}
\title{Function to choose palette}
\usage{
gameR_cols(palette = NULL, reverse = FALSE)
}
\arguments{
\item{palette}{Character name of palette. Either "union" or "pastel".}

\item{reverse}{Logical. Should the palette be reversed? Defaults to FALSE.}
}
\description{
Function to choose palette
}
