% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{gameR_cols}
\alias{gameR_cols}
\title{Choose a gameR palette}
\usage{
gameR_cols(palette = NULL, reverse = FALSE)
}
\arguments{
\item{palette}{Character name of palette. Either "blocks", cups", "cyberpunk",
"fallout", "gris", "rayman", "spirit", "splat", "superbros", or "wow"}

\item{reverse}{Logical. Should the palette be reversed? Defaults to FALSE.}
}
\description{
Choose a gameR palette
}
