% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{gameR_cols}
\alias{gameR_cols}
\title{Choose a gameR palette}
\usage{
gameR_cols(palette = NULL, reverse = FALSE)
}
\arguments{
\item{palette}{Character name of palette. Either
banjo, blocks, border, cowboy, cups, cyberpunk, fallout, gris, kirby, ocarina, okami, p4g, pman, rayman, sonic, spirit, splat, superbros, wow}

\item{reverse}{Logical. Should the palette be reversed? Defaults to FALSE.}
}
\value{
Vector containing a hex color code representation for the chosen
palette
}
\description{
Choose a gameR palette
}
